backend/refine/echo.py
from PIL import Image
import base64
import io

def refine_image(image_base64: str) -> str:
    # Decode base64 → image
    image_bytes = base64.b64decode(image_base64.split(",")[-1])
    image = Image.open(io.BytesIO(image_bytes)).convert("RGB")

    # 🚧 Phase 6.1+ will go here (real model)
    # For now: echo back unchanged

    buffer = io.BytesIO()
    image.save(buffer, format="PNG")
    encoded = base64.b64encode(buffer.getvalue()).decode()

    return f"data:image/png;base64,{encoded}"
