backend/requirements.txt
fastapi
uvicorn
pydantic
pillow
torch
diffusers
transformers
